create CMakeLists
