+++ libserver/smart_contract/.gitignore
@@ -1,0 +1,2 @@
+build
+__pycache__
