+++ libserver/smart_contract/README.md
@@ -1,5 +1,3 @@
-c++调用python3.8
-References https://docs.python.org/3.8/extending/embedding.html#compiling
+## References 
++ c++调用python3.8:https://docs.python.org/3.8/extending/embedding.html#compiling
 
-编译命令 
-g++ -I/usr/include/python3.8/  example_main.cpp -o example_main -L/usr/lib/ -lpython3.8
